requirements.txt

fastapi
uvicorn[standard]
pydantic
langchain-core
langchain-google-genai
langchain-openai
langchain-groq
google-genai
httpx
cryptography
python-multipart
langgraph
langchain-community
pypdf
python-docx
aiokafka
asyncpg
langchain-text-splitters
